move helper methods to booksparams helper and books edgecase helper modules
